backend/pyproject.toml
[project]
name = "vie-backend"
version = "0.1.0"
description = "VIE (Visual Identity Encoder) — adversarial privacy backend"
requires-python = "==3.11.*"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.17",
    "deepface==0.0.93",
    "tf-keras==2.18.0",
    "insightface==0.7.3",
    "onnxruntime==1.20.1",
    "opencv-python-headless==4.10.0.84",
    "pillow==11.0.0",
    "numpy==1.26.4",
    "mediapipe==0.10.18",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vie_backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
